name fixes part 1
Fixes of names of functions and such (part 1, unchecked)

+++ src/main/java/com/uep/wap/repository/AdminActionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminActionRepository extends CrudRepository<AdminAction, Integer> {
+public interface AdminActionRepository extends CrudRepository<AdminAction, Long> {
 }
