+++ src/main/java/com/uep/wap/repository/AnswerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnswerRepository extends CrudRepository<Answer, Integer> {
+public interface AnswerRepository extends CrudRepository<Answer, Long> {
 }
