+++ src/main/java/com/uep/wap/repository/BookmarkRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookmarkRepository extends CrudRepository<Bookmark, Integer> {
+public interface BookmarkRepository extends CrudRepository<Bookmark, Long> {
 }
