+++ src/main/java/com/uep/wap/repository/ModeratorActionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ModeratorActionRepository extends CrudRepository<ModeratorAction, Integer> {
+public interface ModeratorActionRepository extends CrudRepository<ModeratorAction, Long> {
 }
